// frontend/src/client/types.gen.ts
export type AvatarPublic = {
    name: string;
    description?: (string | null);
    id: string;
    owner_id: string;
    image_url: string;
    created_at: string;
    updated_at: string;
};